template/a2a-agents/my-a2a-agent/requirements.txt
a2a-sdk[http-server]>=0.3.23
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
